run.py: Process shaders concurrently.
Now that we've accumulated a non-trivial amount of shaders, running
shader-db takes far too long.  Parsing the shaders in parallel is
embarassingly easy and gives roughly a 4x speed-up on my dual-core
hyperthreaded Ivybridge system.

This may not be the most efficient way to implement the thread pools,
but it seems to work well enough.  Requires Python 3.2.