Skip counting the second pair of an instruction in 16-wide mode.
Generally we want our count to approximate time, and the two SIMD8
instructions are same time as a SIMD16 instruction, except for the
cost to instruction cache pressure.  This pattern happens for MAD and
LRP a lot, and there are other cases, too.  Double-counting these
meant that we basically couldn't use shader-db to validate
optimizations involving those instructions.