run.py: Use shader counts provided by mesa
Instead of counting lines to determine the number of lines in the .out
file, walk the .out file looking for the results that mesa printed into
the out file.

(Note by Ken): This also makes run.py strict - if a shader fails to
compile, it will generate an exception.  This helps prevent people from
commiting corrupt shader files.

Signed-off-by: Dylan Baker <baker.dylan.c@gmail.com>
Acked-by: Kenneth Graunke <kenneth@whitecape.org>