Import Unity 5.0 built-in shaders.
The Unity 3D engine ships with a number of built-in shaders, which will
likely be used by many games.  These were originally written in HLSL,
but translated via a bunch of software.

Thanks to Aras Pranckevičius <aras@unity3d.com> for giving us a copy of
these shaders under the MIT license.