run: Print which shaders were being processed when crashing.
If the driver hits an assertion failure and aborts, the runner will die,
which can be hard to debug.  This patch makes the runner catch SIGABRT
and SIGSEGV and and print out a list of shaders each thread was
processing before it terminates.

 => CRASHED <= while processing these shaders:

    shaders/foo/1936.shader_test
    shaders/foo/9991.shader_test
    shaders/foo/1832.shader_test

v2: Handle SIGSEGV too, since SIGABRT doesn't catch real crashes (such
    as null pointer dereferences).  Call _exit() to avoid continuing.
    Use write() in the signal handler since fputs() is not signal safe.
    According to an Austin Group bug tracker entry, strlen() is signal
    safe, so using it should be fine.