report.py: rework and update for cycle info
Now that we have three separate things we want to measure (instructions,
cycles, and loops), it's impractical to keep adding special code for
changes in each thing. Instead, for each program in before and after we
store a table of measurement -> value, and when reporting we loop over
each measurement and report helped/hurt before reporting the gained/lost
programs.

v2:
- Swap loop count and cycle count to be compatible with older shader-db.
- Fix indentation.
- Report summary information (hurt/helped) last to make it easier to find.
- Squash in second commit to avoid duplicating the list of measurements
due to the above.

Acked-by: Matt Turner <mattst88@gmail.com>
Signed-off-by: Connor Abbott <cwabbott0@gmail.com>