si-report.py: speed up and adjust to debug message reporting, allowing multi-threaded runs
Benefit from recent changes to Mesa master which allows correct reporting
in multi-threaded runs, and generally more robust reporting when comparing
different runs, e.g. that differ in the set of tests that were run.

Parsing also becomes much faster because we (a) do not recompile regexes in
the inner iterations, (b) have fewer lines to parse in total, and (c) are
more selective about when we try to match which regular expression.

While we're at it, parse out S_NOP instructions from the disassembly and
report wait states.

Acked-by: Tom Stellard <thomas.stellard@amd.com>