add special script to run on intel hardware
Intel produces fairly beefy Xeon servers on which it would be nice to be
able to run shader-db to get some results pretty fast. Unfortunately those
don't ship with any intel graphics IP (only ancient Matrox cards).

This new script stubs a bunch of ioctls so that we can run shader-db on
hardware that doesn't have a /dev/dri/renderD128.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@intel.com>