Drop Orbital Explorer shader.
The upstream version of Orbital Explorer doesn't run on Mesa drivers,
as it (unnecessarily) requests a compatibility profile on non-Apple
OSes, and suffers from GLEW bugs.  I sent a pull request two years ago
to port it to libepoxy and get it running on Mesa drivers, but it was
never accepted.

The author eventually emailed me and said that he considers it a
"finished experiment" and said the rendering method (geometry shader
based approach) is inefficient, and he intends to fully rewrite it
someday.

Since this is not representative of typical geometry shaders, and
the project is effectively dead, we may as well remove it.

Reviewed-by: Chad Versace <chadversary@chromium.org>