run: use distinct extension_string variables
The same declaration was being used for two distinct things - a short
EGL and the core GL extensions.

In the former it is a const string, while the latter is a dynamically
managed list of extensions.

Define distinct variables within the specific scope.

v2: use {egl,gl}_extension_string (Eric)

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>