intel_stub: Claim to support I915_PARAM_HAS_CONTEXT_ISOLATION.
Modern kernels support I915_PARAM_HAS_CONTEXT_ISOLATION, which allows
us to whack certain context registers with impunity.  One of those is
INSTPM/CS_DEBUG_MODE2, which controls whether Constant Buffer 0 is
relative to dynamic state base address, or a general GPU address.
If it's relative (the default), we can only push 3 ranges of UBOs.
If it's absolute, we can push 4.  So this affects our shader output.

Follow the behavior of modern kernels and allow us to push all 4 ranges.

Tested-by: Ian Romanick <ian.d.romanick@intel.com>