report: fix '-m inst' case
Right now if the string 'inst' is one of the measurements it is
replaced by 'instructions'. This was done because some drivers were
reporting it as 'inst' and others as 'instructions', and I guess that
also to make the final report easier to read.

But this causes using '-m inst' to fail, as 'instructions' is used to
fill the dictionary. As I found somewhat counterintuitive to assume
that the user should know about that replacement, this patch just adds
the same replacements on other parts of the code.

An alternative would be remove any 'inst' to 'instructions'
replacement, as now drivers output is more consistent. But again, it
is easier to read the report with 'instructions'.

Reviewed-by: Eric Anholt <eric@anholt.net>