run: Limit OMP to the -j threads limit.
The -j argument only worked for the "1" case where OMP is skipped.
Otherwise, we'd spawn the max amount of threads (except for -j 1 where
OMP is skipped) and overflow the current_shader_names array.

Noticed when trying to work around an apparent freedreno bug with too
many contexts on the same screen.

Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>