Name shaders foz/hash/ExecutableName
We're currently only considering one shader per foz file because we
stick stuff in the hash based on the name for which we were only using
the name of the .foz file which meant we only considered a single shader
and it wasn't even guaranteed to be the same shader from both files.
To fix these issues, we instead use .foz/hash/ExecutableName as the name
so we get a complete name for the shader.  This should also let us more
easily identify the shaders in the output.

Reviewed-by: Kenneth Graunke <kenneth@whitecape.org>