Improve fossilize.sh
This commit adds a FOSSILIZE_BUILD_PATH variable which is a bit more
user-friendly than setting VK_LAYER_PATH manually because it's not as
obvious as one would like how to use that environment variable.  It also
appends $(basename $0) to the dump path so we get a nicely named .foz
file.  Finally, we stop adding unneeded directories.